requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0

pydantic>=2.6

chromadb>=0.5

langchain>=0.2
langgraph>=0.2
langchain-community>=0.2
langchain-upstage>=0.1.7

openai>=1.50

langchain-core
langchain-community
langchain-chroma
langsmith
